refactor: improve component organization
