+++ src/components/layout/Footer/Footer.module.scss
@@ -1,4 +1,4 @@
-@use "../../styles/variables" as *;
+@use "../../../styles/variables" as *;
 @use "sass:color";
 
 footer {
@@ -32,13 +32,13 @@ footer {
   width: 100dvw;
   padding: 0 5dvw;
 
-  & > figure {
+  &>figure {
     display: flex;
     flex-direction: column;
     justify-content: center;
     align-items: center;
 
-    & > svg {
+    &>svg {
       width: 100px;
       height: 100px;
     }
@@ -50,11 +50,11 @@ footer {
       align-items: center;
       text-align: center;
 
-      & > figcaption {
+      &>figcaption {
         font-size: 15px;
         white-space: pre-line;
 
-        & > em {
+        &>em {
           font-size: 13px;
 
           strong {
@@ -124,7 +124,7 @@ footer {
   flex-wrap: nowrap;
   gap: 5px;
 
-  & > svg {
+  &>svg {
     flex-shrink: 0;
   }
 }
@@ -141,7 +141,7 @@ footer {
     background-color: $color-base;
   }
 
-  & > svg {
+  &>svg {
     width: 24px;
     height: auto;
     flex-shrink: 0;
@@ -216,7 +216,7 @@ svg {
         padding: 0;
         width: 100%;
 
-        & > figure {
+        &>figure {
           flex-direction: row;
 
           .caption-container {
@@ -231,6 +231,7 @@ svg {
         gap: clamp(5rem, 5vw + 5rem, 20rem);
       }
     }
+
     .footer-simple-divider {
       display: inline;
     }
