+++ src/components/layout/Header/HeaderDesktop.module.scss
@@ -1,4 +1,4 @@
-@use "../../styles/variables" as *;
+@use "../../../styles/variables" as *;
 @use "sass:color";
 
 .header-desktop {
