+++ src/components/layout/Header/HeaderMobile.module.scss
@@ -1,4 +1,4 @@
-@use "../../styles/variables" as *;
+@use "../../../styles/variables" as *;
 
 .header {
   display: flex;
