+++ src/components/layout/Layout/Layout.module.scss
@@ -1,4 +1,4 @@
-@use "../../styles/variables" as *;
+@use "../../../styles/variables" as *;
 
 .layout-wrapper {
   position: relative;
