+++ src/components/layout/index.ts
@@ -1,0 +1,4 @@
+export { Layout } from "./Layout";
+export { Header } from "./Header";
+export { SideMenu } from "./SideMenu";
+export { Footer } from "./Footer";
