+++ src/components/projects/Card/Card.module.scss
