+++ src/components/projects/ImageGallery/ImageGallery.module.scss
