+++ src/components/projects/ImageGallery/ImageGallery.tsx
