+++ src/components/projects/ImageGallery/index.ts
