+++ src/components/projects/index.ts
@@ -1,0 +1,3 @@
+export { Projects } from "./Projects";
+export { Card } from "./Card";
+export { ImageGallery } from "./ImageGallery";
